sorted folder structure and integrated country api
