+++ src/App.js
@@ -1,0 +1,13 @@
+import React, { Fragment } from 'react';
+import Main from './sections/Main';
+
+function App() {
+  return (
+    <Fragment>
+      <h1>GlobeTrotter</h1>
+      <Main />
+    </Fragment>
+  )
+}
+
+export default App;
