+++ src/index.js
@@ -1,14 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Map from './components/Map';
-
-function App() {
-  return (
-    <div>
-      <h1>Globe Trotter!</h1>
-      <Map />
-    </div>
-  )
-}
+import App from './App';
 
 ReactDOM.render(<App />, document.getElementById('root'));
